`fn errno_location`, `static std{out,err}`: Remove from `rav1d` lib and into the binaries.
The binaries aren't important, so it's useful to fully remove these from the important `rav1d` lib.
