+++ tools/input/annexb.rs
@@ -1,10 +1,11 @@
+use crate::compat::errno::errno_location;
+use crate::compat::stdio::stderr;
 use libc::fclose;
 use libc::fopen;
 use libc::fprintf;
 use libc::fread;
 use libc::fseeko;
 use libc::strerror;
-use rav1d::errno_location;
 use rav1d::include::dav1d::data::Dav1dData;
 use rav1d::include::dav1d::headers::Dav1dObuType;
 use rav1d::include::dav1d::headers::DAV1D_OBU_FRAME;
@@ -14,7 +15,6 @@ use rav1d::include::dav1d::headers::DAV1D_OBU_TD;
 use rav1d::include::dav1d::headers::DAV1D_OBU_TILE_GRP;
 use rav1d::src::lib::dav1d_data_create;
 use rav1d::src::lib::dav1d_data_unref;
-use rav1d::stderr;
 use std::cmp;
 use std::ffi::c_char;
 use std::ffi::c_int;
