+++ tools/input/section5.rs
@@ -1,3 +1,5 @@
+use crate::compat::errno::errno_location;
+use crate::compat::stdio::stderr;
 use libc::fclose;
 use libc::feof;
 use libc::fopen;
@@ -5,7 +7,6 @@ use libc::fprintf;
 use libc::fread;
 use libc::fseeko;
 use libc::strerror;
-use rav1d::errno_location;
 use rav1d::include::dav1d::data::Dav1dData;
 use rav1d::include::dav1d::headers::Dav1dObuType;
 use rav1d::include::dav1d::headers::DAV1D_OBU_FRAME;
@@ -15,7 +16,6 @@ use rav1d::include::dav1d::headers::DAV1D_OBU_TD;
 use rav1d::include::dav1d::headers::DAV1D_OBU_TILE_GRP;
 use rav1d::src::lib::dav1d_data_create;
 use rav1d::src::lib::dav1d_data_unref;
-use rav1d::stderr;
 use std::ffi::c_char;
 use std::ffi::c_int;
 use std::ffi::c_uint;
