+++ tools/output/y4m2.rs
@@ -1,10 +1,12 @@
+use crate::compat::errno::errno_location;
+use crate::compat::stdio::stderr;
+use crate::compat::stdio::stdout;
 use libc::fclose;
 use libc::fopen;
 use libc::fprintf;
 use libc::fwrite;
 use libc::strcmp;
 use libc::strerror;
-use rav1d::errno_location;
 use rav1d::include::dav1d::headers::DAV1D_CHR_UNKNOWN;
 use rav1d::include::dav1d::headers::DAV1D_PIXEL_LAYOUT_I400;
 use rav1d::include::dav1d::headers::DAV1D_PIXEL_LAYOUT_I420;
@@ -12,8 +14,6 @@ use rav1d::include::dav1d::headers::DAV1D_PIXEL_LAYOUT_I444;
 use rav1d::include::dav1d::picture::Dav1dPicture;
 use rav1d::include::dav1d::picture::Dav1dPictureParameters;
 use rav1d::src::lib::dav1d_picture_unref;
-use rav1d::stderr;
-use rav1d::stdout;
 use std::ffi::c_char;
 use std::ffi::c_int;
 use std::ffi::c_uint;
